requirements.txt
pyserial 
python-rtmidi
PySimpleGUI==4.70.1
pystray